major update

+++ admin/admin_hapus.php
@@ -2,14 +2,19 @@
 include '../koneksi.php';
 $id = $_GET['id'];
 
-$data = mysqli_query($koneksi, "SELECT * FROM admin WHERE admin_id='$id'");
-$d = mysqli_fetch_assoc($data);
-$foto = $d['admin_foto'];
-if($foto != "" && file_exists("../gambar/user/$foto")) {
-  unlink("../gambar/user/$foto");
+if($id == 1){
+    header("location: settings.php?alert=gagal");
+    exit;
 }
+
+$data = mysqli_query($koneksi, "SELECT admin_foto FROM admin WHERE admin_id='$id'");
+$d = mysqli_fetch_assoc($data);
+if($d['admin_foto'] != "" && file_exists('../gambar/user/'.$d['admin_foto'])){
+    unlink('../gambar/user/'.$d['admin_foto']);
+}
+
 mysqli_query($koneksi, "DELETE FROM admin WHERE admin_id='$id'");
 
-header("location: settings.php?alert=hapus");
+header("location: settings.php?alert=terhapus");
 exit;
 ?>
